Collected Baja Electrical code, both C++ and Sketch.
This repository contains the code for the UBC Baja Electrical sub-team.
Included are previous versions and current versions. The current version
is "XXXXX".

Contained in this version are the main loop in Sketch and the C++
classes, X, Y, and Z. A brief description of each class:
X-
Y-
Z-